+++ src/App.cpp
@@ -7,7 +7,6 @@
  * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
  * Copyright 2016-2018 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
  *
- *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
  *   the Free Software Foundation, either version 3 of the License, or
@@ -30,17 +29,17 @@
 
 #include "api/Api.h"
 #include "App.h"
-#include "Console.h"
+#include "common/Console.h"
+#include "common/log/Log.h"
+#include "common/Platform.h"
 #include "core/Config.h"
 #include "core/Controller.h"
-#include "log/Log.h"
-#include "Platform.h"
 #include "proxy/Proxy.h"
 #include "Summary.h"
 #include "version.h"
 
 #ifndef XMRIG_NO_HTTPD
-#   include "api/Httpd.h"
+#   include "common/api/Httpd.h"
 #endif
 
 
@@ -53,7 +52,7 @@ App::App(int argc, char **argv) :
         return;
     }
 
-    if (!m_controller->config()->background()) {
+    if (!m_controller->config()->isBackground()) {
         m_console = new Console(this);
     }
 
@@ -102,8 +101,8 @@ int App::exec()
     m_httpd = new Httpd(
                 m_controller->config()->apiPort(),
                 m_controller->config()->apiToken(),
-                m_controller->config()->apiIPv6(),
-                m_controller->config()->apiRestricted()
+                m_controller->config()->isApiIPv6(),
+                m_controller->config()->isApiRestricted()
                 );
 
     m_httpd->start();
@@ -113,8 +112,6 @@ int App::exec()
 
     const int r = uv_run(uv_default_loop(), UV_RUN_DEFAULT);
     uv_loop_close(uv_default_loop());
-
-    Platform::release();
 
     return r;
 }
@@ -133,7 +130,7 @@ void App::onConsoleCommand(char command)
     case 'v':
     case 'V':
         m_controller->config()->toggleVerbose();
-        LOG_NOTICE("verbose: %d", m_controller->config()->verbose());
+        LOG_NOTICE("verbose: %d", m_controller->config()->isVerbose());
         break;
 
     case 'h':