+++ src/proxy/LoginRequest.h
@@ -4,8 +4,8 @@
  * Copyright 2014      Lucas Jones <https://github.com/lucasjones>
  * Copyright 2014-2016 Wolf9466    <https://github.com/OhGodAPet>
  * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
- * Copyright 2016-2017 XMRig       <support@xmrig.com>
- *
+ * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
+ * Copyright 2016-2018 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -26,41 +26,38 @@
 
 
 #include <stdint.h>
+#include <string.h>
 
 
 class LoginRequest
 {
 public:
-    enum ClientTypes {
-        OtherClient,
-        XMRigClient,
-        XMRig20Client,   // In versions 2.0.1 and 2.0.2 nicehash not explicitly enabled for donations.
-        XMRigProxyClient // Proxy does not support nicehash for upstream.
-    };
-
     inline LoginRequest() :
-        m_clientType(OtherClient),
         m_agent(nullptr),
         m_login(nullptr),
         m_pass(nullptr),
         m_id(0)
     {}
 
-    LoginRequest(int64_t id, const char *login, const char *pass, const char *agent);
+    inline LoginRequest(int64_t id, const char *login, const char *pass, const char *agent, const char *rigId) :
+        m_agent(agent),
+        m_login(login),
+        m_pass(pass),
+        m_rigId(rigId),
+        m_id(id)
+    {}
 
     inline const char *agent() const      { return m_agent; }
     inline const char *login() const      { return m_login; }
     inline const char *pass() const       { return m_pass; }
+    inline const char *rigId() const      { return m_rigId && strlen(m_rigId) > 0 ? m_rigId : m_login; }
     inline int64_t id() const             { return m_id; }
-    inline ClientTypes clientType() const { return m_clientType; }
 
 private:
-    ClientTypes detectClient() const;
-
-    ClientTypes m_clientType;
     const char *m_agent;
     const char *m_login;
     const char *m_pass;
+    const char *m_rigId;
     const int64_t m_id;
 };