+++ src/proxy/Stats.cpp
@@ -23,9 +23,9 @@
 
 
 #include "api/Api.h"
+#include "common/net/SubmitResult.h"
 #include "Counters.h"
 #include "interfaces/ISplitter.h"
-#include "net/SubmitResult.h"
 #include "proxy/events/AcceptEvent.h"
 #include "proxy/Stats.h"