+++ src/proxy/events/LoginEvent.h
@@ -4,8 +4,8 @@
  * Copyright 2014      Lucas Jones <https://github.com/lucasjones>
  * Copyright 2014-2016 Wolf9466    <https://github.com/OhGodAPet>
  * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
- * Copyright 2016-2017 XMRig       <support@xmrig.com>
- *
+ * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
+ * Copyright 2016-2018 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -25,6 +25,7 @@
 #define __LOGINEVENT_H__
 
 
+#include "common/crypto/Algorithm.h"
 #include "proxy/events/MinerEvent.h"
 #include "proxy/LoginRequest.h"
 
@@ -32,19 +33,21 @@
 class LoginEvent : public MinerEvent
 {
 public:
-    static inline LoginEvent *create(Miner *miner, int64_t id, const char *login, const char *pass, const char *agent)
+    static inline LoginEvent *create(Miner *miner, int64_t id, const char *login, const char *pass, const char *agent, const char *rigId, const xmrig::Algorithms &algorithms)
     {
-        return new (m_buf) LoginEvent(miner, id, login, pass, agent);
+        return new (m_buf) LoginEvent(miner, id, login, pass, agent, rigId, algorithms);
     }
 
 
+    const xmrig::Algorithms &algorithms;
     LoginRequest request;
 
 
 protected:
-    inline LoginEvent(Miner *miner, int64_t id, const char *login, const char *pass, const char *agent)
+    inline LoginEvent(Miner *miner, int64_t id, const char *login, const char *pass, const char *agent, const char *rigId, const xmrig::Algorithms &algorithms)
         : MinerEvent(LoginType, miner),
-          request(id, login, pass, agent)
+          algorithms(algorithms),
+          request(id, login, pass, agent, rigId)
     {}
 };