+++ src/proxy/splitters/nicehash/NonceMapper.cpp
@@ -27,18 +27,17 @@
 #include <string.h>
 
 
+#include "common/log/Log.h"
+#include "common/net/Client.h"
+#include "common/net/strategies/FailoverStrategy.h"
+#include "common/net/strategies/SinglePoolStrategy.h"
 #include "core/Config.h"
 #include "core/Controller.h"
-#include "log/Log.h"
-#include "net/Client.h"
-#include "net/strategies/FailoverStrategy.h"
-#include "net/strategies/SinglePoolStrategy.h"
-#include "net/Url.h"
+#include "net/JobResult.h"
 #include "proxy/Counters.h"
 #include "proxy/Error.h"
 #include "proxy/events/AcceptEvent.h"
 #include "proxy/events/SubmitEvent.h"
-#include "proxy/JobResult.h"
 #include "proxy/Miner.h"
 #include "proxy/splitters/nicehash/NonceMapper.h"
 #include "proxy/splitters/nicehash/NonceStorage.h"
@@ -99,7 +98,7 @@ void NonceMapper::gc()
 }
 
 
-void NonceMapper::reload(const std::vector<Url*> &pools)
+void NonceMapper::reload(const std::vector<Pool> &pools)
 {
     delete m_pending;
 
@@ -128,6 +127,10 @@ void NonceMapper::submit(SubmitEvent *event)
 
     if (!m_storage->isValidJobId(event->request.jobId)) {
         return event->reject(Error::InvalidJobId);
+    }
+
+    if (event->request.algorithm.isValid() && event->request.algorithm != m_storage->job().algorithm()) {
+        return event->reject(Error::IncorrectAlgorithm);
     }
 
     JobResult req = event->request;
@@ -172,7 +175,7 @@ void NonceMapper::onActive(IStrategy *strategy, Client *client)
         m_pending  = nullptr;
     }
 
-    if (m_controller->config()->verbose()) {
+    if (m_controller->config()->isVerbose()) {
         LOG_INFO(isColors() ? "#%03u \x1B[01;37muse pool \x1B[01;36m%s:%d \x1B[01;30m%s" : "#%03u use pool %s:%d %s",
                  m_id, client->host(), client->port(), client->ip());
     }
@@ -181,12 +184,16 @@ void NonceMapper::onActive(IStrategy *strategy, Client *client)
 
 void NonceMapper::onJob(IStrategy *strategy, Client *client, const Job &job)
 {
-    if (m_controller->config()->verbose()) {
-        LOG_INFO(isColors() ? "#%03u \x1B[01;35mnew job\x1B[0m from \x1B[01;37m%s:%d\x1B[0m diff \x1B[01;37m%d" : "#%03u new job from %s:%d diff %d",
-                 m_id, client->host(), client->port(), job.diff());
+    if (m_donate) {
+        if (m_donate->isActive() && client->id() != -1 && !m_donate->reschedule()) {
+            m_donate->save(client, job);
+            return;
+        }
+
+        m_donate->setAlgo(job.algorithm());
     }
 
-    m_storage->setJob(job);
+    setJob(client->host(), client->port(), job);
 }
 
 
@@ -221,17 +228,20 @@ void NonceMapper::onResultAccepted(IStrategy *strategy, Client *client, const Su
 
 bool NonceMapper::isColors() const
 {
-    return m_controller->config()->colors();
+    return m_controller->config()->isColors();
 }
 
 
-IStrategy *NonceMapper::createStrategy(const std::vector<Url*> &pools)
+IStrategy *NonceMapper::createStrategy(const std::vector<Pool> &pools)
 {
+    const int retryPause = m_controller->config()->retryPause();
+    const int retries    = m_controller->config()->retries();
+
     if (pools.size() > 1) {
-        return new FailoverStrategy(pools, m_controller->config()->retryPause(), m_controller->config()->retries(), this);
+        return new FailoverStrategy(pools, retryPause, retries, this);
     }
 
-    return new SinglePoolStrategy(pools.front(), m_controller->config()->retryPause(), this);
+    return new SinglePoolStrategy(pools.front(), retryPause, retries, this);
 }
 
 
@@ -257,6 +267,18 @@ void NonceMapper::connect()
 {
     m_suspended = 0;
     m_strategy->connect();
+}
+
+
+void NonceMapper::setJob(const char *host, int port, const Job &job)
+{
+    if (m_controller->config()->isVerbose()) {
+        LOG_INFO(isColors() ? "#%03u " MAGENTA_BOLD("new job") " from " WHITE_BOLD("%s:%d") " diff " WHITE_BOLD("%d") " algo " WHITE_BOLD("%s")
+                            : "#%03u new job from %s:%d diff %d algo %s",
+                 m_id, host, port, job.diff(), job.algorithm().shortName());
+    }
+
+    m_storage->setJob(job);
 }