+++ src/proxy/splitters/simple/SimpleMapper.cpp
@@ -27,18 +27,17 @@
 #include <string.h>
 
 
+#include "common/log/Log.h"
+#include "common/net/Client.h"
+#include "common/net/strategies/FailoverStrategy.h"
+#include "common/net/strategies/SinglePoolStrategy.h"
 #include "core/Config.h"
 #include "core/Controller.h"
-#include "log/Log.h"
-#include "net/Client.h"
-#include "net/strategies/FailoverStrategy.h"
-#include "net/strategies/SinglePoolStrategy.h"
-#include "net/Url.h"
+#include "net/JobResult.h"
 #include "proxy/Counters.h"
 #include "proxy/Error.h"
 #include "proxy/events/AcceptEvent.h"
 #include "proxy/events/SubmitEvent.h"
-#include "proxy/JobResult.h"
 #include "proxy/Miner.h"
 #include "proxy/splitters/simple/SimpleMapper.h"
 
@@ -71,7 +70,7 @@ void SimpleMapper::add(Miner *miner, const LoginRequest &request)
 }
 
 
-void SimpleMapper::reload(const std::vector<Url*> &pools)
+void SimpleMapper::reload(const std::vector<Pool> &pools)
 {
     delete m_pending;
 
@@ -95,6 +94,20 @@ void SimpleMapper::reuse(Miner *miner, const LoginRequest &request)
 }
 
 
+void SimpleMapper::stop()
+{
+    m_strategy->stop();
+
+    if (m_pending) {
+        m_pending->stop();
+    }
+
+    if (m_donate) {
+        m_donate->stop();
+    }
+}
+
+
 void SimpleMapper::submit(SubmitEvent *event)
 {
     if (!isActive()) {
@@ -103,6 +116,10 @@ void SimpleMapper::submit(SubmitEvent *event)
 
     if (!isValidJobId(event->request.jobId)) {
         return event->reject(Error::InvalidJobId);
+    }
+
+    if (event->request.algorithm.isValid() && event->request.algorithm != m_job.algorithm()) {
+        return event->reject(Error::IncorrectAlgorithm);
     }
 
     JobResult req = event->request;
@@ -141,7 +158,7 @@ void SimpleMapper::onActive(IStrategy *strategy, Client *client)
         m_pending  = nullptr;
     }
 
-    if (m_controller->config()->verbose()) {
+    if (m_controller->config()->isVerbose()) {
         LOG_INFO(isColors() ? "#%03u \x1B[01;37muse pool \x1B[01;36m%s:%d \x1B[01;30m%s" : "#%03u use pool %s:%d %s",
                  m_id, client->host(), client->port(), client->ip());
     }
@@ -150,9 +167,10 @@ void SimpleMapper::onActive(IStrategy *strategy, Client *client)
 
 void SimpleMapper::onJob(IStrategy *strategy, Client *client, const Job &job)
 {
-    if (m_controller->config()->verbose()) {
-        LOG_INFO(isColors() ? "#%03u \x1B[01;35mnew job\x1B[0m from \x1B[01;37m%s:%d\x1B[0m diff \x1B[01;37m%d" : "#%03u new job from %s:%d diff %d",
-                 m_id, client->host(), client->port(), job.diff());
+    if (m_controller->config()->isVerbose()) {
+        LOG_INFO(isColors() ? "#%03u " MAGENTA_BOLD("new job") " from " WHITE_BOLD("%s:%d") " diff " WHITE_BOLD("%d") " algo " WHITE_BOLD("%s")
+                            : "#%03u new job from %s:%d diff %d algo %s",
+                 m_id, client->host(), client->port(), job.diff(), job.algorithm().shortName());
     }
     setJob(job);
 }
@@ -185,7 +203,7 @@ void SimpleMapper::onResultAccepted(IStrategy *strategy, Client *client, const S
 
 bool SimpleMapper::isColors() const
 {
-    return m_controller->config()->colors();
+    return m_controller->config()->isColors();
 }
 
 
@@ -204,13 +222,16 @@ bool SimpleMapper::isValidJobId(const xmrig::Id &id) const
 }
 
 
-IStrategy *SimpleMapper::createStrategy(const std::vector<Url*> &pools)
+IStrategy *SimpleMapper::createStrategy(const std::vector<Pool> &pools)
 {
+    const int retryPause = m_controller->config()->retryPause();
+    const int retries    = m_controller->config()->retries();
+
     if (pools.size() > 1) {
-        return new FailoverStrategy(pools, m_controller->config()->retryPause(), m_controller->config()->retries(), this);
+        return new FailoverStrategy(pools, retryPause, retries, this);
     }
 
-    return new SinglePoolStrategy(pools.front(), m_controller->config()->retryPause(), this);
+    return new SinglePoolStrategy(pools.front(), retryPause, retries, this);
 }