+++ src/proxy/workers/Workers.cpp
@@ -4,8 +4,8 @@
  * Copyright 2014      Lucas Jones <https://github.com/lucasjones>
  * Copyright 2014-2016 Wolf9466    <https://github.com/OhGodAPet>
  * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
- * Copyright 2016-2018 XMRig       <support@xmrig.com>
- *
+ * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
+ * Copyright 2016-2018 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -25,9 +25,9 @@
 #include <inttypes.h>
 
 
+#include "common/log/Log.h"
 #include "core/Config.h"
 #include "core/Controller.h"
-#include "log/Log.h"
 #include "proxy/events/AcceptEvent.h"
 #include "proxy/events/CloseEvent.h"
 #include "proxy/events/LoginEvent.h"
@@ -38,7 +38,7 @@
 
 
 Workers::Workers(xmrig::Controller *controller) :
-    m_enabled(controller->config()->workers()),
+    m_enabled(controller->config()->isWorkers()),
     m_controller(controller)
 {
 }
@@ -60,7 +60,7 @@ void Workers::printWorkers()
     char workerName[24];
     size_t size = 0;
 
-    Log::i()->text(m_controller->config()->colors() ? "\x1B[01;37m%-23s | %-15s | %-5s | %-8s | %-3s | %11s | %11s |" : "%-23s | %-15s | %-5s | %-8s | %-3s | %11s | %11s |",
+    Log::i()->text(m_controller->config()->isColors() ? "\x1B[01;37m%-23s | %-15s | %-5s | %-8s | %-3s | %11s | %11s |" : "%-23s | %-15s | %-5s | %-8s | %-3s | %11s | %11s |",
                    "WORKER NAME", "LAST IP", "COUNT", "ACCEPTED", "REJ", "10 MINUTES", "24 HOURS");
 
     for (const Worker &worker : m_workers) {
@@ -178,7 +178,7 @@ void Workers::accept(const AcceptEvent *event)
 
 void Workers::login(const LoginEvent *event)
 {
-    const std::string name(event->request.login());
+    const std::string name(event->request.rigId());
 
     if (m_map.count(name) == 0) {
         const size_t id = m_workers.size();